CMakeLists.txt: Introduce ENABLE_PACKAGE_SOURCE option and rework bc_compute_full_version()
ENABLE_PACKAGE_SOURCE option makes the creation of 'package_source' custom target optional,
allowing to build with CMake < 3.11 if it is disabled (it is by default).

bc_compute_full_version() is now able to compute a version number for
hotfix commits that have not been tagged as a new 'patch' release. Such
commits are now considered as alpha versions for the next 'patch' release.